Develope

+++ final-project/src/pages/Panel/products/Components/Table/table.product.component.js
@@ -4,17 +4,36 @@ import Styles from './table.product.module.css';
 import {IMAGE_URL} from '../../../../../configs/image.url';
 import { Icon } from '@iconify/react';
 import { useDispatch } from 'react-redux';
-
+import { useState } from 'react';
 import { loadCurrentItem } from '../../../../../redux/Shopping/shopping-actions';
-
+import { confirmAlert } from 'react-confirm-alert'; 
+import 'react-confirm-alert/src/react-confirm-alert.css'; 
+import Modal from '../../../../../Components/Modal/modal.page'
 
 const TableProductComponent = (props) => {
     const dispatch = useDispatch();
-    const handleDelete = () => 
+    
+   const  submit = () => {
+        confirmAlert({
+            
+          title: 'تایید حذف',
+          message: 'آیا از حذف این محصول اطمینان دارید؟',
+          buttons: [
             {
+              label: 'بله',
+              onClick: () =>  {
                 console.log(props.id)
-                props.deleteFunc(props.id)
+                 props.deleteFunc(props.id)
+                 window.location.reload()
             }
+            },
+            {
+              label: 'خیر',
+              onClick: () => ""
+            }
+          ]
+        });
+      };
 
             function handleClick(props) {
                 dispatch(loadCurrentItem(props));
@@ -29,13 +48,14 @@ const TableProductComponent = (props) => {
 
 
         <div className= {Styles.tableHead}>
+
             
             
            <div className={Styles.Btns}>
                <button 
                onClick = {() => handleClick(props)}
                >ویرایش</button>
-               <button onClick={handleDelete}> 
+               <button onClick={submit}> 
                <Icon className={Styles.deleteItem}   icon="mdi:delete" color="#ee2d40" width="30" height="30" />
                </button>
            </div>
